qtaskdist_worker/requirements.txt
amazon-braket-default-simulator==1.32.0
amazon-braket-schemas==1.26.1
amazon-braket-sdk==1.102.10
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
asyncpg==0.30.0
attrs==25.4.0
backoff==2.2.1
backports.entry-points-selectable==1.3.0
boltons==25.0.0
boto3==1.40.61
boto3-stubs==1.40.61
botocore==1.40.61
botocore-stubs==1.40.61
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cirq-core==1.6.1
cirq-ionq==1.6.1
cloudpickle==2.2.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dill==0.4.0
duet==0.2.9
fonttools==4.60.1
graphviz==0.21
greenlet==3.2.4
ibm-cloud-sdk-core==3.24.2
ibm-platform-services==0.70.0
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
lark==1.3.1
llvmlite==0.45.1
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
numba==0.62.1
numpy==2.3.4
openpulse==1.0.1
openqasm3==1.0.1
opt_einsum==3.4.0
oqpy==0.3.9
packaging==25.0
pandas==2.3.3
pillow==12.0.0
ply==3.11
psutil==7.1.2
pycparser==2.23
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.10.1
pyparsing==3.2.5
pyqasm==1.0.0
pyqir==0.10.9
pyspnego==0.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytket==2.10.2
pytket-braket==0.44.0
pytz==2025.2
qbraid==0.10.0
qbraid-core==0.1.45
qbraid-qir==0.4.0
qiskit==2.2.2
qiskit-ibm-runtime==0.41.1
qiskit-qasm3-import==0.6.0
qwasm==1.0.1
redis==7.0.1
requests==2.32.5
requests_ntlm==1.3.0
rustworkx==0.17.1
s3transfer==0.14.0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
stevedore==5.5.0
sympy==1.14.0
tqdm==4.67.1
types-awscrt==0.28.2
types-s3transfer==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
